Restructured classes Both ProcessingNode and RecorderNode now inherit from the common ConfigHandlerNode
