+++ src/processing_node/processing_node/recorder_node.py
@@ -1,5 +1,6 @@
 import csv
 import sys
+import os
 
 from asyncio import Future
 from io import StringIO
@@ -7,8 +8,7 @@ from io import StringIO
 import ament_index_python
 import rclpy
 
-from processing_node.processing_node import ProcessingNode
-from processing_node.ml_deploy_library import *
+from processing_node.config_handler_node import ConfigHandlerNode
 
 def get_config_file_path(filename):
     # Get the directory of the package
@@ -19,18 +19,13 @@ def get_config_file_path(filename):
 
     return config_file_path
 
-class RecorderNode(ProcessingNode):
+class RecorderNode(ConfigHandlerNode):
     """
     ROS2 node that sets up input subscribers according to a config and writes
     the there published data into an outfile
 
     Attributes:
         outfile (StringIO): File-like Object to write the data into
-
-        supported_message_types (dict): dict that is filled with imported message types
-
-        aggregated_input_data (dict):  dict that is filled with data gathered from
-        the set up subscribers to be written into the outfile
 
         recording_done (Future): Future object that is set to done when the number of
         input points set during construction or by altering of the respective ROS-parameter
@@ -43,35 +38,18 @@ class RecorderNode(ProcessingNode):
 
         Args:
             config_path (str): Path to the config specifiying the inputs and imports
-
             outfile (StringIO): File-like object to write the data into
-
             number_of_input_poins (int): Number of points to be written into the outfile.
+            frequency (float): Dictates how often the execute method is called
         """
-        rclpy.node.Node.__init__(self, "recorder_node")
 
-        self._config = load_config(config_path)
-
-        config_is_valid, error_message = check_if_config_is_valid(self._config)
-        if not config_is_valid:
-            raise ValueError(error_message)
+        super().__init__(config_path, frequency, "recorder_node")
 
         self.outfile = outfile
 
-        self.supported_message_types = import_needed_modules(self._config)
-        self.aggregated_input_data = dict.fromkeys(self._config['Inputs'].keys(), [])
-
-        input_topic_dict = map_input_and_output_names_to_topics(self._config)[0]
-        self.set_up_subscriptions(input_topic_dict)
+        self.recording_done = Future()
 
         self.declare_parameter("number_of_input_points", number_of_input_points)
-
-        # Set model timer period
-        timer_period = 1.0 / frequency
-
-        self.timer = self.create_timer(timer_period, self.execute)
-
-        self.recording_done = Future()
 
     def execute(self):
         """
